Set precision of output

+++ linux/main.cpp
@@ -102,7 +102,7 @@ void speedometerFunction(){
     time(&currentTime_2);
 	printf("currentTime_2 %s", ctime(&currentTime_2));
     //timeDifferenceSeconds = difftime(currentTime_2,currentTime_1);
-	timeDifferenceSeconds = currentTime_2 - currentTime_1;
+	timeDifferenceSeconds = double(currentTime_2 - currentTime_1);
 	cout << "timeDifferenceSeconds:" << timeDifferenceSeconds << endl;
     speedometerReadingCalculation(timeDifferenceSeconds);
 	cout << "MPH:" << milesPerHour << setprecision(5) << endl;
